Make camera-coord tracklets, clean up (theta,r,theta2).
Make the tracklets that are used into camera coordinates, and the ones
for plotting purposes into world coordinates. This should help
prevent "cheating" and also cleans up the main script.

The camera relation is now returned in (theta1,r,theta2) format when
calculated, and it's now is in line with the definition. This also
cleans up main a little bit.